Limpieza de código - Anda todo
Embebido -> Comentarios para los defines a cambiar por tiempos, y cambio de nombre en algunos defines. Limpieza de código en general.

Registrador -> Se juntó a los testers y al registrador en el mismo archivo, y ahora los 3 son funciones (loops infinitos) a llamar dentro del main, según se elija registrar o hacer una prueba. Se limpió el código en general, cambiando detalles y ordenando comentarios, renombrando defines, etc...

SQL Server -> Se agregó la columna IdReg, y se cambió el nombre de MINS_ON por el de MINS_DIFF, dado que la columna MINS_ON no existe más en la base (en la tabla registros), y ésta, se calcula en la vista como la diferencia de minutos entre las columnas LAST_DT y INIT_DT.

+++ Embebido/HilosLibertad_V1/Core/Inc/types.h
@@ -6,12 +6,12 @@
     #include <stddef.h>
 
 
-	#define MILISEG                         1
-    #define SEG                             1000*MILISEG
-    #define MIN                             60*SEG
+	#define MILISEG                         ((int) 1)
+    #define SEG                             ((int) 1000 * MILISEG)
+    #define MIN                             ((int) 60 * SEG)
 
 
-    // struct byte (bits, nibbles)
+    // struct byte (como bits o nibbles)
     typedef struct byte_bits{
         uint8_t bit0 : 1;
         uint8_t bit1 : 1;
@@ -33,7 +33,7 @@
         byte_bits bits;
     } byte;
 
-    // struct board (bits, nibbles, bytes)
+    // struct board (como bits, nibbles o bytes)
     typedef struct board_bits{
         uint8_t bit00 : 1;
         uint8_t bit01 : 1;
@@ -70,7 +70,7 @@
         board_bits bits;
     } board;
     
-    // struct sample (bits, nibbles, bytes, boards)
+    // struct sample (como bits, nibbles, bytes o boards)
     typedef struct sample_bits{
         uint8_t bit00 : 1;
         uint8_t bit01 : 1;
